+++ src/Arbitrary.hs
@@ -60,6 +60,7 @@ import Action.Implementation
 import Config
 import EventLog
 import Frontend.Core
+import Frontend.Fragment.Comment
 import Frontend.Page
 import Frontend.Prelude (set, (^.), (.~), ppShow, review, view, join)
 import LifeCycle
