+++ src/Frontend/Fragment/Comment.hs
@@ -9,15 +9,7 @@
 
 {-# OPTIONS_GHC -Werror -Wall #-}
 
--- | ...
---
--- We provide data types @Page...@ even if there is an application type already.  Example: For
--- 'Idea', we define 'PageIdea'.  This has at least two benefits:
---
--- - page types should always be defined here to avoid orphans;
--- - we can add additional information (like author name if we only have an author's id) and thus
---   avoid making page rendering effectful.
-module Frontend.Page.Comment
+module Frontend.Fragment.Comment
 where
 
 import qualified Generics.SOP as SOP
